etc/requirements-LATEST.txt
Cython==0.20.2
Django==1.5.8
FontTools==2.4
FullContact.py==0.0.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.5.0
PyYAML==3.11
Pygments==1.6
SQLAlchemy==0.9.6
South==1.0
Sphinx==1.2.2
anyjson==0.3.3
argcomplete==0.8.1
argh==0.25.0
argparse==1.2.1
args==0.1.0
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.3.2
boto==2.30.0
bpython==0.13
clint==0.3.7
couchdbkit==0.6.5
cssselect==0.9.1
decorator==3.4.0
django-admin-bootstrapped==1.6.5
django-appconf==0.6
django-autoslug==1.7.2
django-countries==2.1.2
django-delegate==0.3.0
django-docfield-couchdb==0.2.4
django-extensions==1.3.3
django-haystack==2.1.1-dev
django-imagekit==3.2.1
django-jquery-widgets==0.1.2
django-memcached==0.1.2
django-modeldict==1.4.1
django-multifilefield==1.0
django-nose==1.2
django-postgrespool==0.2.4
django-storages==1.1.8
django-supervisor==0.3.2
django-tablib==3.0.2
django-tagging==0.3.2
django-twemoir==0.2.1
django-uuidfield==0.5.0
docutils==0.11
eventlet==0.14.0
fake-useragent==0.0.5
fontaine==1.1.5
gnureadline==6.3.3
greenlet==0.4.2
guess-language==0.2
gunicorn==19.0.0
h5py==2.4.0a0
hiredis==0.1.4
honcho==0.5.0
http-parser==0.8.3
httpie==0.8.0
httplib2==0.9
humanize==0.5
ijson==1.1
imread==0.3.3-git
instakit==0.2.3
ipython==2.1.0
jsonfield==0.9.20
lxml==3.3.5
mahotas==1.1.1
matplotlib==1.3.1
meld3==1.0.0
nltk==2.0.4
nodeenv==0.9.6
nose==1.3.3
numexpr==2.4
numpy==1.8.1
oauth2==1.5.211
oauthlib==0.6.3
pandas==0.14.0
pathtools==0.1.2
phyjnius==1.2.2
pilkit==1.1.6
psycopg2==2.5.3
pycurl==7.19.3.1
pyicu-praxa==1.7
pylibmc==1.3.0
pylire==0.3.8
pymc==2.3.3
pyparsing==2.0.2
pyquery==1.2.8
pysolr==3.1.0
pysqlite==2.6.3
pystache==0.5.4
python-dateutil==2.2
python-magic==0.4.6
python-twitter==1.3.1
pytools==2014.3
pytz==2014.4
redis==2.10.1
requests==2.3.0
requests-oauthlib==0.4.1
restkit==4.2.2
scikit-image==0.10.1
scikit-learn==0.14.1
scipy==0.14.0
setproctitle==1.1.8
simplejson==3.5.2
six==1.7.3
socketpool==0.5.3
static==1.0.2
superlance==0.9
supervisor==3.0
tables==3.1.1
tablib==0.10.0
tornado==3.2.2
twitter-text-python==1.0.1
uWSGI==2.0.6
ujson==1.33
urllib3==1.8.3
urlstring==2.4.0
viron==0.4.8
watchdog==0.8.0
wsgiref==0.1.2
xerox==0.3.1
yolk==0.4.3
